refactor(researcher): Mover track_metrics a capa de aplicación
- Movido decorador track_metrics a application/decorators/
