+++ app/agents/researcher/domain/entities/__init__.py
@@ -1,5 +1,13 @@
 from .metrics_data import MetricsData
 from .search_engine import SearchEngine
 from .query_validation import QueryValidation
+from .research_status import ResearchStatus
+from .research_state_schema import ResearchStateSchema
 
-__all__ = ['MetricsData', 'SearchEngine', 'QueryValidation']
+__all__ = [
+    'MetricsData',
+    'SearchEngine',
+    'QueryValidation',
+    'ResearchStatus',
+    'ResearchStateSchema'
+]
